rename: modify component name to solve duplicate export
