+++ apps/web/src/features/post/filterPostCategoryAccordion/ui/FilterPostCategoryAccordion.css.ts
