+++ apps/web/src/features/post/index.ts
@@ -1,4 +1,4 @@
 export { useTogglePostLike } from './togglePostLike/model/useTogglePostLike';
 export { SelectViewMode } from './selectViewMode/ui';
 export { TogglePostLike } from './togglePostLike/ui/TogglePostLike';
-export { FilterCategory } from './filterCategory/ui/FilterCategory';
+export { FilterPostCategoryAccordion } from './filterPostCategoryAccordion/ui/FilterPostCategoryAccordion';
