+++ apps/web/src/features/spot/filterSpotCategory/ui/FilterSpotCategory.css.ts
